♻️ [Refactoring]: #135 GraphicsStateStackの循環依存を解消
